+++ compiler/src/main/kotlin/com/squareup/metro/extensions/scoped/ContributesMultibindingScopedMetroExtension.kt
@@ -78,7 +78,7 @@ public class ContributesMultibindingScopedMetroExtension(private val session: Fi
       MetroContributionExtension.Contribution(
         supertype = metroContributionSymbol.defaultType(),
         replaces = emptyList(),
-        originClassId = parentSymbol.classId,
+        originClassId = contributionInterfaceClassId,
       )
     }
   }